TBSONBinaryProtocol fix for mismatched array type error
it turns out that i was not fully reading the bytes requested off the
wire because InputStream.read(byteArray, offset, length) is not
guaranteed to actually read length bytes even if it has those bytes
available. Solution is to loop until the total requested bytes are read.

this was hard to figure out because the behavior of InputStream.read is
platform dependent. It worked fine in my OSX development environment,
but not in Linux

(sapling split of 1a90633b7e2ba95f0c136f16f58af000baa435e9)