Move buildgen core into Fsq.io
Foursquare has been using buildgen to maintain
our pants BUILD files. We have been asked to
open source it over the years and we finally
found the time to refactor it to the point
where it could be usable for other repos.

This is the engine, we hope to push python
and java implementations soon.

(sapling split of 796602f898a9a5aa68f2516f609cbeed114fa600)