Inline BuildFileManipulator and bring bg subsystem into core.
The subsystem is a logical part of the core - it is only imported
by core and core won't work without it.

This also inlines BuildFileManipulator from upstream pants.
I am looking into distributing the buildgen modules as
contrib sdists and I don't know if there is a mechanism
to import classes from other contrib modules - much of the
contrib handling is private. But this is most useful
as a part of the buildgen core module, now that it
will be oss.

I also modified the way buildgen changes are printed to
terminal - the output should be the same as ever, but
now it respects the --no-colors flag. That change is in
the 'write' method of the imported BuildFileManipulator.

(sapling split of 94f93cb57a94ff7cc89d3d6cf0a0013a5f9a3401)