Add a catch for jvm addresses that aren't in the build graph.
An OSS consumer was getting errors because dep was None - which
happens when the build graph tries to resolve an address that isnt
injected int the build graph.

This was not is Fsqio, so they were not using pom-resolve and
therefore must have been using ivy.

If they had been using pom-resolve, this will be caught long
before buildgen. It is unclear how it got triggered at all - I was
unable to repro using Ivy either.

Further, I believe that Pants should and does have checks for this.

But the fact remains that a buildgen consumer of 1.0.0 pants
was hitting that state. Adding the used addresses to 3rdparty allowed
buildgen progress to continue, so here is a constant-time check
to provide a useful error should it happen to anyone else.

(sapling split of fae9a131251a02972b18feee20b964ff039519f4)