Split up the thrift files used by Spindle tests
The generated class files were just too big- they were
causing builds to OOM during GC with multi-second pauses
for Humongous allocation GC occuring even on laptops.

This was a blocker for Travis CI for Fsq.io, which we need
if we are ever to consume patches from outside contributors.

The refactor is just a re-org that splits them up into smaller
modules. There was a bit of guess work here - but I think that
in general the layout of the tests has improved. There should be
no functional changes outside of the renames.

I also got rid of all the ._ imports that used to be
used in these files and the com.foursquare strings within
the tests. Both of these have long-standing asanas.

The one target that is split that might make more sense as
one is the test/thrift/io/fsq/spindle/codegen/runtime/test and
runtime/struct. But it needed to be split - I still saw OOMs
with those targets as one single target.

(sapling split of 7820dd5a8ce90ff5acca51eb24c11f267b2fcb8d)