Remove the pom_cache_dir from the fingerprint
This was going to make CI runs have a fp that conflicted
with the versions elsewhere. It was non-trivial to
work around it.

It appears to be unneeded, in that pom-resolve already verified
the location of every symlink. Recreating the symlink farm every
run isn't ideal but it isn't any worse than what it is already
doing.

(sapling split of 807f64560378fb9a9faf62164fbfdba45e5283cd)