Adjust local_override_versions
This eases the overhead of iterating on a fork of a jar library.
Since we only allow one version of a given jar, this allows the
override to match on the (org, name) key.

This will NOT invalidate on an override change. Since the symlink
farm is rebuilt for every run, new overrides will be reflected
I added a println as a sanity check - the logging of pom-resolve
needs a rework.

There is the possibilty of mysterious errors if the change to the jar
involves adding transitive dependencies. Those cases will require more
hands-on work.

My original work was trying to add the ability to resolve jars from the
local `~/.m2` repo. This would be the robust solution and would respect
transitive deps. Resolving from the local m2 is "what you would expect".

But even when I get that working there will still be
a place for local_override_versions, since you simply won't always want
to pay the price of a ful resolve for every iteration.

(sapling split of 05eb406c6a3cf900231bb784b5d6220808deb361)