Invalidation correctness improvements for webpack plugin
This code had already landed but was reverted due to problems
with post-build resource processing.

I was able to repro the original issue and also verify that it
no longer triggers when run locally. I believe that corresponds
with the errors we were seeing in CI.

Build details:
    * removed the workunits since the invoked upstream plugin manages them
    * Added a WebpackDistribution synthetic target to represent the resolved
      distribution.
      * Certain invalidations were not propogating downstream,
          * for example bumping the webpack_resolve task_version did not invalidate
            webpack codegen, and a couple other edge cases.

This reintroduces fingerprint changes we have wanted to land for awhile,
the codegen/resolve fingerprinting was off.

(sapling split of 217a98e8bcea8d97ef2e5a5e2fe7cd0ffb7f8247)