Raise actionable errors when import parsing fails.
When doing refactors, there are often erroneous BUILD file references
left in the codebase. This improves the error in one such circumstance,
where a sources field points to a non-existent file.

The message was really hard to act on,  because it was a constructed path,
the string it reports on likely doesn't exist anywhere in the codebase.
It was easy to determine that the file indeed didn't exist - not so easy
to find what was pointing to it.

The triggering example was a 'register.py' listed as a source file
under src/python/fsqio/pants/spindle.

* Old error (not helpful when you have hundreds of changed files):
  <trace only references the import parse>

  Exception message: Failed to parse python source: src/python/fsqio/pants/spindle/register.py
  Error opening source: src/python/fsqio/pants/spindle/register.py
  [Errno 2] No such file or directory: u'src/python/fsqio/pants/spindle/register.py'

* New Error:

      <snip>
      File "/Users/mateo/dev/foursquare.web/src/python/fsqio/pants/buildgen/python/buildgen_python.py", line 299, in buildgen_target
      """.format(e, source_candidate, target.address.spec)

  Exception message:

  Failed to parse: src/python/fsqio/pants/spindle/tasks/register.py
  [Errno 2] No such file or directory: u'src/python/fsqio/pants/spindle/tasks/register.py'

              Source file: src/python/fsqio/pants/spindle/tasks/register.py
              Referenced by: src/python/fsqio/pants/spindle/tasks:register

(sapling split of de8327fa7c51b43085ffc1e98019ca317b6022f0)