Enable 'test' and 'repl' for webpack
This enables the support for webpack_tests
and ./pants repl - good news from this commit.

* repl works on any webpack target
* test works on webpack_test targets.

We have a node_test that can be run by changing
to webpack_test - but the test is blocking
because it launches in the browser - more akin
to a `webpack.debug` or something.

Try it out!

In the futurem I am considering generating the package.json
from BUILD files. Just using the hand generated one was fine,
but I may try that out in the future. Not using the BUILD file
model has resulted in a lot of overrides.

(sapling split of 8ddbef90ed20b21af5b13aa3e87abab6157b6496)