Remove the upstream NodePaths product requirement
With ivy resolution install, this causes a task goal cycle.
NodePaths is registered in 'resolve' but we use WebPack as 'gen'.

So that is a cycle, since gen _has_ to come before resolve. In
reality, this probably makes it a mistake to have had Webpack
use SimpleCodegen. But it was able to give us the rest of the
Jvm pipeline for free, so probably worth it.

The solution used here was just to make a new product and force the
webpack goal and all tasks to happen earlier in the pipeline.

(sapling split of 4032b332960bf02bc4909844f098ecee16b2600d)