Set output of junit to FAILURE_INLY by default.
Our fs wrapper already defaults to this, but it is best to
ensure that Pants does this the same way. It is almost always
what is wanted.

This also removed some redundant config in the ini and the
wrapper.

(sapling split of f516d4e9b5a682de08341934767c01e764c6bd36)