run Futures.groupedCollect with a window rather than batching
Summary:
Currently, we wait on the slowest Future in each batch to complete
before moving on to the next batch, even though we could actually
begin running some of its Futures and still remain under the
concurrency limit. This change adds a wrapper to the given function
that pulls the next param to process from a queue and start its
computation once the current Future is satisfied.

(sapling split of 5550875ff7f4e863836e4358eacfc7fb1add13b4)