Update upkeep to no longer clobber global variables
The scripts have all been updated to no longer rely on relative
paths, there is a standard concept of a BUILD_ROOT and it must be
set for the tasks to run.

There is some extra protections on some tasks, in cases
where the task is thought to be especically destructive.

(sapling split of 4990e7c868559e5bbf6598031269ed39d85a2bb8)