Move the env.sh task to a subdir of the upkeep namespace.
This really is a special file (conditionally sourced depending on
which namespaces are available in the checkout).

Moved to a silo'd area to make clear it is not a task or an "action"
but a library file that is being handled by the machinery and should
not be executed or git blamed, like ever.

(sapling split of 20b59fa5fd06014047a576c11d9e6ce909462a13)