Add scalac jars to 3rdparty
So this is a bit of guesswork, and complicated.

The buildgen scalac plugins in this commit are a specialized
target, the scalac_plugin. They compiled internally, but failed
when ripped out into fsqio.

The buildgen targets have source, but with no deps. It imports
scalac and scala-reflect, but was failing to find them on the
compile classpath when compiled within fsq.io.

The target was not different and when I added the scalac_plugin_jars
options like the internal plugins, the issue still existed.

The reason appears to be because we have additional scalac plugin
targets internally, and they happen to have scalac and scala-reflect
jars as dependencies in their pom.xml. Pants appears to generate
a hermetic-esque environment, where it constructs a classpath
contructed of the scalac_plugin source files and anything passed
to the scalac_plugin_jars argument.

So to fix, we needed to pass scala-reflect and scalac as scalac_plugin_jars.
But that didn't quite work either, since only scala-reflect is actuallyu
resolved by pom-resolve, which inserts itself in-between the ivy resolve
of the pants tools (including scalac) and the contruction of any
compile-classpath. So adding [ //:scalac, 3rdparty:scala-reflect] only
added the reflect jar to the classpath.

The solution you see here was to add scalac to pom-resolve. I also removed the
duplicated scala jar_library targets from //BUILD.opensource - those meant
that there were two different scala-reflect jar libraries, for example. We
just never noticed because one was resolved by ivy and another by pom-resolve.

I added scalac and scala-reflect as dependencies to the buildgen plugins,
in the hopes that the next time they are built, they would transitively
end up on the classpath.

But even that is not great - internally we appear to have a true bootstrap
problem, in that the compile-time dependencies of the plugin source code
are being brought in by forcing a compile-time dependency on the compiled
plugin itself :|

(sapling split of 0146332b9638a01e439e283f0247ac8eca11f9ce)