Force a clean
The incremental compiler is somehow munging the analysis
files. It is unclear to me exactly what is wrong.

At least one guess is that we didn't order a clean with the pants upgrade.
We changed the cache key, but that may not be respected for the
incremental compiler.

The cache key needs to be mixed into the task cache_key - it now
occurs to me that it may not be.

(sapling split of 20f3deaa8b21ac4978f5c0c8058b7cae4c7a87f2)