MongoClientManager for the new CRUD api
This is modeled after lift's MongoDB singleton. However, you will note
that 1. it does not live in lift and 2. it is not a singleton. This
removes the need to proxy all our mongo operations through lift and
prevents the need for a lift fork to support the async client and new
api.

This also serves as an abstraction layer around the otherwise disjoint
blocking and async mongo clients. However, even with most of the
business logic pulled out you will notice that the two client managers
are almost byte-for-byte identical. This is regretable, but unavoidable
without resorting to structural types and reflection.

(sapling split of f65db335e7bb5318f2b5c9c837b84469b840aa0d)