add BlockingMongoClientAdapter
Here we have the client adapter for the blocking api. This includes
a wrapper class to carry around query results, which allows us to
share most of the adapter logic with the async version. Implicits
are also included to allow seamless conversion between wrapped and
unwrapped values.

Note the TODO around BlockingResult, ideally that class would be an
extension of AnyVal so it does not incur a runtime cost. However, this
gets us into erasure issues with the type abstracted wrapEmptyResult
method. I don't think this is too much of an issue in practice, but
certainly something we should work around if possible.

(sapling split of 102aabd2508edcc7d2c88e4c0c4ece37f861c40c)