implement count in MongoClientAdapter
This should give a more concrete overview of what implementing the
various adapter methods will look like: the bulk of the logic is
shared in the base adapter and the two subclasses are only
responsible for the boilerplate of actually dispatching the request.

Note we are forced here to add an implicit conversion for async
callbacks returning a java.lang.Long, it is possible we will need
a few more of those as we go.

Tests will be incoming.

(sapling split of 0d97e79c773d65952d0e9a8b5cbbbf1b9e91ec3b)