add new QueryExecutor
This is also more or less a port of the old QueryExecutor to the
new api, the key difference now being that we have to dispatch to
the client adapter for empty query results since the return type
is now abstracted away.

Note that there is only a single QueryExecutor class for use with
async and blocking clients, as well as spindle and lift, which is
configured via the class parameters (eventually this will also
include the model SerDe helpers, but that was not necessary here).
This then, is the final unification point between the various apis
in the new backend.

(sapling split of 50788e156784661b265504eb2798143d0238b8fc)