add LiftMongoCollectionFactory
This is another mostly-port from existing code, this time
LiftDBCollectionFactory. Note this class is type parameterized by
the mongo collections it handles, and with the MongoClientManager
interface functions both with the async and blocking client.

The relevant commands are included in the class comment to
construct a functioning QueryExecutor capable of running count
queries.

(sapling split of bc687628397e5a3f345ee81b2e67eb2ce51e8719)