pull SalatDAO.scala into the twofishes tree
Salat does not have an official release that supports version 3.X of
casbash and the mongo-java-driver. I have opened a pull request for
that here: https://github.com/salat/salat/pull/188, however it is
unknown when that change will actually go through.

In the meantime, we need a solution that also works for our opensource
users. To that end I am pulling SalatDAO.scala into our repo under the
io.fsq namespace, as this is the only file we actually need to fork in
order to run the indexer successfully. It is copied as is here, and
our patch will be applied in a follow-up commit.

(sapling split of 67b0837fc3a4a71e67ff8c79de5748a5a2e4fe0b)