basic harness for thread-safe mongo tests
The idea here is that we are essentially just mangling db names
in order to give each test its own view of the database and
prevent data races.

Users are required to implement an init function, which in the
simplest case just records pre-existing client connections in
the client managers. I would love to open source our connection
setup code to flesh this out a bit more, but that is need of
a cleanup first. For now, the Rogue tests will have to suffice
as examples.

(sapling split of be1ec9955c4815afe33b90b71d06c9dcc994714f)