Bump cache key to optimistically wipe a poisoned spindle cache
We migrated a method out of a spindle dependency and into fsq.io.
But even after a clean, we are seeing the old package namespace
on the compile classpath.

The target is the original location of the moved ThriftResusingSerializer.
The easiest culprit to blame is in generated spindle code. That code is
not safe for reuse and is almost always where we find cache poisoning.
I have found out where the issue in that is, and will try to solve
permanently sometime in the near future.

(sapling split of 45f7f97439f51296227a41d40817bcc2ef8ac26f)
(sapling split of 0605ad47739716e6d09862921b944a39c7127fbc)
(sapling split of d71f9555c0ccdf178ba8425b3db6e399551c4b00)