add a new TrivialORMQueryTest
This is another refresh of existing code for the new rogue and
mongo apis. This will be the primary location for test cases in
the new rogue world and I will be adding more as query
functionality is implemented.

This is also the first consumer of the new MongoTest helper for
thread-safe mongo tests and gives a good example of how the
setup for them should work. I expect us internally (and any
external Rogue users) to actually extend MongoTest with a more
specialized helper that abstracts the initClientManagers step
away from individual tests, such that its setup is only
happening in one place.

(sapling split of 409936ee58ca02f0926abd8acb1b9710d7d70a47)