Add RemoteSource target to RpmBuilder pipeline.
This allows us to host versioned bundles and pull them
down using Pants binary_utils pipeline.

This is basically meant for sources that can be or are
already versioned. A good example is a bundle of source
code or a precompiled release.

When creating RPMs, iterating was painful because it had to
redownload the remote source every run. When using this target
it instead caches it under the pants supportdir, and the
downloads persist even through a clean.

We rely on the binary_util namespacing format, urls
can be declared in the [binaries] scope.

Example usage will land in a followup, where I am converting
a couple SPECS to the rpmbuilder and leveraging this.

I am not wedded to the RemoteSources name, it is alraeady
overloaded - HostedSource or VersionedBundle?

lets workshop it and/or never think about it again

(sapling split of 956a3e65cf6fb44b861f88fc84819d7b3904447b)