implement TrivialORM count tests
Tests are included for both the async and blocking client. Note
that I also had to do some finagling with the async client here,
as the default codec registry for that does not include DBObject
which we currently need. See the inline comment for more
information.

(sapling split of f835f7ee286aedb6cd17c71ec050bdaed3f359d0)