add test for MongoTest harness
This commit adds HarnessTest, a test designed to prove the
correctness of the name mangling performed by MongoTest in order
to provide thread safety.

Strictly speaking, this test is unneccessary; if the test harness
is broken our other tests will likely fail as a result. However,
this provides a way to separate out failures in the test library
itself from standard test failures due to other reasons and is
much simpler to debug.

(sapling split of a99736237da461992dcc7bbfa1fac56987f200c4)