RpmSpecs can depend on other RpmSpec targets.
Simple enough - this adds all RPMs created by RpmSpecs to a product.
If an RpmSpecTarget depends on another, this copies the created RPM
to the build workspace and passes that location to yum during the
BuildRequirements install phase.

This runs the RpmSpec targets through the invalidation framework in order
to get dependency ordering and to create a workdir under pants.d to store
the created RPMs. But build_rpm is called for every target, if it is
invalidated or not. That is not necessary in my mind, but the Payload would
need to be vetted and it is probably better to just handle that through the
DockerPlatform.

(sapling split of 5b67a73523e46bd50f4edd24c76c02c850878194)