rogue v2 - implement fetch
This is unfortunately more boiler platey than it otherwise would be
due to the async and blocking FindIterables not sharing any common
interface. That said, this still accomplishes the goal of holding
all business logic in the base client adapter, and hiding
boilerplate is a core purpose of an ORM layer anyway.

(sapling split of 9e324145a89b48f1c7820804b05cf060c66cd409)