test concurrency fix-up
- Run some previously serial tests in parallel, now that we are using
WriteConcern.W1 for mongo writes in tests
- Move serial annotations in-line to the class level, or method level
if possible. This is to make concurrency opt-out instead of opt-in as
much as possible.
- A few obvious one line fixes to tests that were previously flaky
when run concurrently.

The remaining serial tests (which are most of them) fall into a few
categories:

- mongo tests with bulk deletes/finds/counts/etc. These will be fixed
once we are switched to the new mongo test harness
- tests that rely on execution timing or timeouts
- tests which manipulate the filesystem

(sapling split of 585d1a642bd8f79ce4170e983b1e2c31b88bfea4)