Break out the spindle-stubs into a subclass.
This
* removes the references to the stubs from spindle_gen

* allows us to override the `product_type`, which was causing
the new task to be scheduled as a part of every compile.

* Gives the stubs a unique workdir, to avoid a situation of
populating the generated spindle cache with stubs and breaking the
world.

The biggest difference is the stubs path. I undid the change that moved
Pants working dir to being under the stubs_path, since that breaks
some of Pants guarantees. It now uses the regular .pants.d workspace
and afterwards copies those files to the stub path.

That means we could consider adding this task to a CI sometime and
then shipping the stubs in the artifact cache. Not today though.

(sapling split of c312e87c52b28bb9f6b4fe4bd7cfb6b977c82125)