Make webpack_resolve task fingerprint stable between local/CI.
The webpack-resolve is per-target, which leads to a lot of duplication.
As we consider adding more targets, the resolution time quickly outgrows
the plugins ability to scale.

The hermetic resolution can be especially troublesome because
since it is per-target, it is paid by every engineer, for every change.

* Why remote caching was disabled for webpack targets

NPM uses dangling symlinks as a part of its failure-tolerant ecosystem -
a resolve that fails still generates a symlink to where the lib _would_
have gone, had the resolve succeeded. The js generation is considered
a success as long as transpiling is successful.

The Pants run itself was actually successful, but trouble hit when
Pants was attempting to create the artifact in order to
populate the cache. With the dangling symlinks, the tar invocation
was failing. And because the artifact creation and population is done
in a background process, we never saw the tar failure in the logs or
in the console - in fact Pants would often print SUCCESS and then immediately
followup with FAILED once the artifact cache callback returned.

During the creation of this patch, I saw fp churn between the CI machines
and the local developer boxes, and that had to be solved prior to shipping.
The CI invocations were properly creating artifacts, but to a different
cache_key from local runs of the same checkout.

At first I thought it was erroneously including the cache read/write options,
since that cache subsystem is registered under that task. But it turned out,
obvious in retrospect, that the dereferenced symlinks option was
only set in the CI config.

Dereferncing symlinks or not changes the complexion of the artifact, and
not matching between the invocations was properly munging the cache_key.

(sapling split of 4b334cf3e909f5c06a79999add515c096e002666)