Upgrade node RPMs to 6.9.5 and upgrade Pants plugin.
This RPM also builds a new npm version of 3.10.10, which
fixes the bug that has been biting us of not respecting
the optional modules, and therefore failing CIs when a
transitive module is platform dependent.

I manually invalidated the webpack resolve task since it
did not do so automatically when I bumped the node version.
That is a regression, I had wired up the node version to
the fingerprint, so I do not know why that is required.
I will follow up with that separately, for now this is a
complete solution for the webpack consumers.

(sapling split of 9fb4e90139a0464acf2de7d9012ef785cfe92ae8)