new rogue - implement remove
This comes with another restriction on the Document type being a subtype
of Bson. Again, this should be fine as all built in document classes in
the driver subclass Bson, but if need be we can always remove these
restrictions and add some additional view methods to
MongoCollectionFactory.

(sapling split of 7e9cb9c47bb9b98b8cb2853ef5f620c0a82963b6)