new rogue - add transformResult
Also make wrapEmptyResult a call-by-name method that can also process
exceptions.

This is intended to be an abstract interface to future.transform, in
order to allow transformations of async results. A follow up commit
makes use of this to fix logging in the async client adapter.

(sapling split of af149fdff135599096f733d39b3ed85536e55311)