override equals and hashCode on MongoIdentifier
Similar to what lift does for ConnectionIdentifier, this makes any
use in maps/comparisons/etc based solely on the string name. This is
actually necessary in the case of LiftMongoCollectionFactory, where
the alternative is a cast that may not succeed depending on how meta
records are configured.

(sapling split of 813d41d845d27a79bc0a4fa5670d782c3e1a868b)