switch rogue's logCounter helper off varargs
There used to be an issue with the implementation of varargs making a full
copy of the given arguments. I could not find any information on if that
was ever fixed in the language, but the fix to simply change the interface
given we haven't had an official rogue release since I added it.

(sapling split of 192e30094f7e0f9f159b672904c972dd5ad2984a)