pants: use pants' scala_jar helper instead of per-file sjar helper (#266)
Pants has a way to append the proper Scala version info to the name of
jars that contain Scala code. This was not the case when we first
wrote the `sjar` helper. Switch to using the Pants `scala_jar` syntax.
(sapling split of 883b4cca15b7d01c2c462b2a05615d7d49c51729)