Add a check for the current file before deleting (#709)
This is not the most deliberate way to do this, but it is
the safest. The recent change made it possible that `current_path`
was unset. Passing the unquoted param to 'rm -rf' means that
some very destructive consequences were imaginable.

The trigger for noticing this was that it would run the
delete logic even if called on a non-existing path. The thing
that saved us from bad consequences was that the captured result
of the find_upkeep_file captures the error string on failure,
which in itself is not what it was designed to do, but possibly
also saved some pain in this instance.
(sapling split of 331c71bc8a624696d7748e36a5cc10d3aacf65f7)