Delete the custom JDK install for Fsq.io CI. (#744)
When the TravisCI was configured for fsq.io, travis did not have
a solution for configuring JDK for python projects.

It is not clear they have that answer today, but they did
make JDK8 the default. We have to kill the custom logic
because Oracle has killed the hack we were using to download
the JDK tarball from their download site.
(sapling split of cc8459c99ec8491c588e3c01a1b4b8e4b46063c5)